枚举修改提交

+++ domain/src/main/java/com/cdkhd/npc/entity/Notification.java
@@ -64,6 +64,6 @@ public class Notification extends BaseDomain {
             2、区上
 	 */
    	@Column(name = "level" )
-	private Integer level;
+	private Byte level;
 
 }
